requirements.txt
Flask
psycopg2
werkzeug
python-dotenv
